working dump nodes helper function
