+++ src/functions/node_helper.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use PhpParser\Node;
 use PhpParser\PrettyPrinter\Standard;
+use Symfony\Component\Console\Output\OutputInterface;
 
 if (! function_exists('print_node')) {
     /**
@@ -19,5 +20,25 @@ if (! function_exists('print_node')) {
             $printedContent = $standard->prettyPrint([$node]);
             var_dump($printedContent);
         }
+    }
+}
+
+if (! function_exists('dump_node')) {
+    /**
+     * @param Node|Node[] $node
+     */
+    function dump_node(Node|array $node)
+    {
+        $styler = \Illuminate\Container\Container::getInstance()
+            ->make(\Rector\Console\Style\SymfonyStyleFactory::class)
+            ->create();
+
+        // we turn up the verbosity so it's visible in tests overriding the
+        // default which is to be quite during tests
+        $styler->setVerbosity(OutputInterface::VERBOSITY_VERBOSE);
+
+        $styler->newLine();
+
+        (new \Rector\Util\PrintNodes($styler))->outputNodes($node);
     }
 }
